comments service and exception improvement
