+++ grails-app/services/org/broadinstitute/orsp/QueryService.groovy
@@ -11,6 +11,9 @@ import org.broadinstitute.orsp.webservice.PaginationParams
 import org.grails.plugins.web.taglib.ApplicationTagLib
 import org.hibernate.Criteria
 import org.hibernate.FetchMode
+import org.hibernate.HibernateException
+import org.hibernate.MappingException
+import org.hibernate.NonUniqueResultException
 import org.hibernate.SQLQuery
 import org.hibernate.SessionFactory
 import org.hibernate.transform.Transformers
@@ -1383,10 +1386,19 @@ class QueryService implements Status {
         final String query =
                 ' select * from comment where project_key = :issueId'
         final SQLQuery sqlQuery = session.createSQLQuery(query)
-        final results = sqlQuery.with {
-            addEntity(Comment)
-            setString('issueId', issueId)
-            list()
+        Collection<Comment> results = null
+        try {
+            results = sqlQuery.with {
+                addEntity(Comment)
+                setString('issueId', issueId)
+                list()
+            }
+        } catch(NonUniqueResultException e) {
+            log.error("There is more than one matching result when trying to get comments for IssueId: ${issueId}.", e)
+        } catch(MappingException e) {
+            log.error("The given entity name could not be resolved, when trying to get comments for issueId: ${issueId}.", e)
+        } catch(HibernateException e) {
+            log.error("An error has occurred when trying to get comments for issueId: ${issueId}.", e)
         }
         results
     }
